refactor(core): extract 8 computation methods from ROIDashboardService into ROIDashboardService.compute
Moves calculateTimeSaved, groupBySkill, calculateWeeklyTrend, getWeekStart,
computeStakeholderROIOnTheFly, computeUserMetrics, computeSkillMetrics,
and computeDailyMetrics to a dedicated module-level compute file.

ROIDashboardService.ts: 497 → 349 lines (under 500-line governance limit)
ROIDashboardService.compute.ts: 187 lines (new, private helpers extracted)

All 25 ROIDashboardService tests pass. TypeScript strict check clean.

SMI-5201
